pyproject.toml
[tool.poetry]
name = "bksys_transaction_ms"
version = "0.1.0"
description = ""
authors = ["malo.girard <malo.girard@jellysmack.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
prometheus-fastapi-instrumentator = "^6.1.0"
python-logging-loki = "^0.3.1"
cryptography = "^41.0.7"
aiomysql = "^0.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
